[FIX] last commit changed resolution of minimap, which is not optimal, instead, this one change the size of chunks in ft_memcpy

+++ Cub3d_bonus/src/utils/ft_memcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvigny <jvigny@student.42.fr>              +#+  +:+       +#+        */
+/*   By: qthierry <qthierry@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 10:22:12 by jvigny            #+#    #+#             */
-/*   Updated: 2023/08/25 20:37:29 by jvigny           ###   ########.fr       */
+/*   Updated: 2023/10/12 20:17:39 by qthierry         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	register size_t	i;
-	register size_t	size;
+	register u_int32_t	i;
+	register u_int32_t	size;
 
-	size = sizeof(size_t);
+	size = sizeof(u_int32_t);
 	i = 0;
 	if (n > size)
 	{
 		while (i < n - size)
 		{
-			*(size_t *)(dest + i) = *(size_t *)(src + i);
+			*(u_int32_t *)(dest + i) = *(u_int32_t *)(src + i);
 			i += size;
 		}
 	}
